add base class for common truncation logic
