+++ lenskit/lenskit/metrics/ranking/_base.py
@@ -1,4 +1,4 @@
-from abc import abstractmethod
+from abc import ABC, abstractmethod
 from typing import Protocol
 
 from lenskit.data import ItemList
@@ -6,8 +6,40 @@ from lenskit.data import ItemList
 
 class RankingMetric(Protocol):
     """
-    Base class / protocol for ranking metrics.
+    Protocol implemented by ranking metrics.
     """
 
     @abstractmethod
     def __call__(self, recs: ItemList, test: ItemList) -> float: ...
+
+
+class RankingMetricBase(ABC, RankingMetric):
+    """
+    Base class for most ranking metrics, implementing a ``k`` parameter for
+    truncation.
+
+    Args:
+        k:
+            Specify the length cutoff for rankings. Rankings longer than this
+            will be truncated prior to measurement.
+    """
+
+    k: int | None = None
+    "The maximum length of rankings to consider."
+
+    def __init__(self, k: int | None = None):
+        if k is not None and k < 0:
+            raise ValueError("k must be positive or None")
+        self.k = k
+
+    def truncate(self, items: ItemList):
+        """
+        Truncate an item list if it is longer than :attr:`k`.
+        """
+        if self.k is not None:
+            if not items.ordered:
+                raise ValueError("top-k filtering requires ordered list")
+            if len(items) > self.k:
+                return items[: self.k]
+
+        return items
